requirements.txt
streamlit
streamlit-mic-recorder
langchain
ctransformers
transformers
pyyaml
librosa
pypdfium2
InstructorEmbedding
huggingface
torch
langchain_community
langchain_groq
sentence_transformers
faiss-cpu
pypdf
accelerate
soundfile